reth/pyproject.toml
[tool.poetry]
name = "reth"
version = "0.0.8"
description = ""
authors = ["reth <sosp20212021@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
numba = "^0.51.2"
numpy = "^1.19.2"
readerwriterlock = "^1.0.7"
gym = "^0.17.3"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
